updated advanced task

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,7 +8,7 @@
 int main(void)
 {
 	int c, boolean, boolean2;
-	long int n1, n2, fn, fn2, n11, n22;
+	int n1, n2, fn, fn2, n11, n22;
 
 	n1 = 1;
 	n2 = 2;
